Added forgot password screen and rebased it on develop
