--- domain/onboarding/src/main/java/com/example/oboarding_domain/forgot_password/use_case/ForgotPasswordUseCase.kt
@@ -1,5 +1,0 @@
-package com.example.oboarding_domain.forgot_password.use_case
-
-data class ForgotPasswordUseCase(
-    val sendResetLink: SendResetLink,
-)
