+++ ui/onboarding/src/main/java/com/amalitech/onboarding/di/onboardingPresentationModule.kt
@@ -1,8 +1,8 @@
 package com.amalitech.onboarding.di
 
 import com.amalitech.onboarding.OnboardingViewModel
+import com.amalitech.onboarding.forgot_password.ForgotPasswordViewModel
 import com.amalitech.onboarding.login.LoginViewModel
-import com.amalitech.onboarding.reset_password.ResetPasswordViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -13,8 +13,7 @@ val onboardingPresentationModule = module {
     single {
         LoginViewModel(get())
     }
-
-    viewModel {
-        ResetPasswordViewModel(get())
+    single {
+        ForgotPasswordViewModel(get())
     }
 }
